AMP-30858 - Create Excel Importer for Burkina

+++ amp/WEB-INF/src/org/digijava/module/admin/action/DataImporter.java
@@ -267,7 +267,7 @@ public class DataImporter extends Action {
         Transaction commitment  = new Transaction();
         commitment.setCurrency(currencyId);
         commitment.setAdjustment_type(adjType);
-        commitment.setTransaction_amount(Long.parseLong(String.valueOf(amount)));
+        commitment.setTransaction_amount(Double.parseDouble(String.valueOf(amount)));
         commitment.setTransaction_date(fundingDate);
         funding.getCommitments().add(commitment);
         funding.getDisbursements().add(commitment);
